# openapi.yaml
openapi: 3.0.3
info:
  title: Desarrolla360 Empresas API
  version: "1.0.0"
  description: >
    Internal API surface for the Desarrolla360 B2B training portal. Every route here
    is consumed either by the portal's own frontend, by Vercel Cron, or by the
    Tutor LMS / WordPress bridge plugin — none of it is a public product API.
servers:
  - url: https://empresas.desarrolla360.com
    description: Production
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: authjs.session-token
      description: >
        NextAuth v5 session cookie, set on login. The endpoint additionally
        requires the session's role to match what's listed below.
    cronSecret:
      type: http
      scheme: bearer
      description: >
        `Authorization: Bearer <CRON_SECRET>`. Accepts `CRON_SECRET`,
        `BACKGROUND_SYNC_SECRET`, or `SYNC_JOBS_SECRET`. An `x-sync-secret`
        header with the same value also works.
    webhookSignature:
      type: apiKey
      in: header
      name: x-d360-webhook-signature
      description: >
        HMAC-SHA256 of `{timestamp}.{rawBody}` using `BRIDGE_WEBHOOK_SECRET`,
        hex-encoded. Sent alongside `x-d360-webhook-timestamp` (Unix seconds,
        must be within 10 minutes of the server clock).
    portalKey:
      type: apiKey
      in: header
      name: X-D360-Portal-Key
      description: >
        Shared secret configured in the plugin's settings page (`WP_BRIDGE_PORTAL_KEY`
        on the portal side). `Authorization: Bearer <key>` with the same value also
        works. A logged-in WordPress admin (`manage_options`) is accepted too.
...
paths:
...
  /api/certificates/{id}/dc3:
    get:
      tags: [Certificates]
      summary: Download a certificate's official DC-3 PDF
      description: >
        Generates (or reuses a cached) DC-3 PDF for one certificate. Accessible to the
        certificate's own employee, the HR of the owning company, or a SUPERADMIN.
      security: [{ sessionCookie: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: download
          in: query
          schema: { type: string, enum: ["1"] }
          description: "Set to `1` to force `Content-Disposition: attachment` instead of `inline`."
      responses:
        "200":
          description: The DC-3 PDF.
          content:
            application/pdf:
              schema: { type: string, format: binary }
        "400":
          description: Invalid certificate id.
        "401":
          description: Not authenticated.
        "403":
          description: Authenticated, but not the certificate owner, their HR, or a SUPERADMIN.
        "404":
          description: Certificate not found.
        "422":
          description: The course is missing DC-3 metadata (duration, thematic area, instructor, etc).
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: string }
                  missing:
                    type: array
                    items: { type: string }
...
  /api/upload/instructor-signature:
    post:
      tags: [Uploads]
      summary: Upload an instructor's signature image
      description: Flattens to a white background PNG and stores it in Vercel Blob.
      security: [{ sessionCookie: [SUPERADMIN] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, wpCourseId]
              properties:
                file:
                  type: string
                  format: binary
                  description: PNG, JPEG, or WebP, up to 2 MB.
                wpCourseId:
                  type: integer
                nombre:
                  type: string
                  description: Optional display name used to derive the stored filename.
      responses:
        "200":
          description: Signature stored.
          content:
            application/json:
              schema:
                type: object
                properties:
                  url: { type: string }
        "400":
          description: Missing/invalid file, invalid wpCourseId, wrong format, or file too large.
        "401":
          description: Not authenticated or not a SUPERADMIN.
...
  /api/internal/sync/employee-learning:
    get:
      tags: [Cron]
      summary: Sync a batch of stale employees' learning data from Tutor LMS
      description: >
        Accepts either a cron secret or a SUPERADMIN session. Vercel Cron calls this
        via GET every 15 minutes; POST is kept for manual/external schedulers.
      security:
        - cronSecret: []
        - sessionCookie: [SUPERADMIN]
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 25 }
      responses:
        "200":
          description: Batch sync result.
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean, enum: [true] }
                  sync_interval_ms: { type: integer }
                  scanned: { type: integer }
                  queued: { type: integer }
                  results:
                    type: array
                    items: { type: object }
        "401":
          description: Missing/invalid cron secret and not a SUPERADMIN session.
    post:
      tags: [Cron]
      summary: Sync a batch of stale employees' learning data (manual trigger)
      security:
        - cronSecret: []
        - sessionCookie: [SUPERADMIN]
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 25 }
      responses:
        "200":
          description: Same shape as GET.
        "401":
          description: Missing/invalid cron secret and not a SUPERADMIN session.
...